+++ mg_custom_nodes/ZhiHui6_zhihui_nodes_comfyui_20260409/Nodes/ImgSwitch/ImageSwitchDualMode.py
@@ -9,7 +9,7 @@ class ImageSwitchDualMode:
             "required": {
                 "mode": (["manual", "auto"], {"default": "auto"}),
                 "select_image": ([str(i) for i in range(1, 1024)], {"default": "1"}),
-                "inputcount": ("INT", {"default": 2, "min": 2, "max": 1000, "step": 1}),
+                "inputcount": ("INT", {"default": 2, "min": 2, "max": 12, "step": 1}),
                 "image1_note": ("STRING", {"multiline": False, "default": ""}),
                 "image2_note": ("STRING", {"multiline": False, "default": ""}),
             },